added key derivation from user password instead of hardcoded key in the source code. We store the IV per session in the meta file, so we can generate the key again when the user logs in.
Perhaps we should add checking for new users that their password is "secure" since the key now partially depend on it.

the nice thing is that we reencrypt with new key (not that important) and new IV (semantic security) after each use of the program, so an adversary that simply looks at the files will not know if we still have the same things stored. Albeit a user might always tore the same passwords forever, so..