Removed logging of username, since a user can accidentially confuse user and password fields and result in failed logon which then logs the password! auch
Also finally fixed the reencryption hazzle. We now reencrypt in all flows. The code is VERY stateful, so it should be refactored into a state machine instead of all these ifs. The state machine should transtition and know how to handle a form closing. Right now we set class-wide variables all over the place.. its gross and hard to reason about. But it works!