StoredPasswordIO class implemented for Json format, used for import export.
Goal is to use this both for db and import export, and outphase the Nvs stuff (dont necc delete it, just dont use it).

What we need to do / fix is that iterations are low, they need to be 500k as well for key derivation, only 30k currently.

..which means that the derivation halts the program when you save. So we need the same task logic that we currently have for Nvs, that we start deriving as soon as we can, and restart.

But that stuff should be internal to the IO classes. start the task in the constructor, and we call new constructors all over the place, but do it correctly to restart it. or expose a method.

This will really clean up the storedpasswordrepostiory class, since it will no longer have to keep track of keys and salt. salt will be stored alongside data.