a lot of refactoring. We are trying to get the decrypting and encrypting IO to control the keys themselves, and write the salt to the underlying files.
This seems to work now, but we still need to clean up A LOT in StoredPasswordRepository, all the references and usages to keys.

Also, do we need to update the salt in the user anymore? maybe for writing the user? hm...

We also need to remove all the salt for keys from the user security data.

Once we are done, everything will be much nicer. Then all the crypto is in the db layer, and the repository does not need to keep track of it, and the user does not need to store the salt.. who ever came up with that in the first place.